+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * In query error detection and feedback through detailed `DevoClientException` 
 ### Removed
  * Parameter `key` removed from `devo.sender.lookup.Lookup.send_data_line`. The `key` parameter pointed to the value in the `fields` list that was the key of the lookup item. When the value appeared several times in `fields`, the one expected to be the key cannot be identified. This parameter was set as deprecated since version 3.4.0 (2020-08-06) 
+### Changed
+* API query timeout by default set to 300 seconds, instead of 30
+* Retries timeout following Exponential backoff algorithm. Default retry delay base set to 5 seconds
+* Retry delay configurable through `retry_delay` instead of `timeout` parameter
 
 ## [4.0.3] - 2022-09-12
 ### Fix